Update zshrc vimrc And Add nvm, rvm installer

+++ bin/setup.d/zsh.sh
@@ -5,9 +5,8 @@ DIR=$(cd $(dirname $0)/../..;pwd)
 
 # Instal Zshrc
 
-if [ -f "$HOME/.zshrc" ] 
-then
-	echo "Skip For .zshrc (File Exists)"
-else
-	ln -s "$DIR/_zshrc.sh" "$HOME/.zshrc"
-fi
+[ -f "$HOME/.zshrc" ]               || ln -s "$DIR/_zshrc.sh" "$HOME/.zshrc"
+[ -f "$HOME/.zprofile" ]            || ln -s "$DIR/_zprofile.sh" "$HOME/.zprofile"
+[ -d "$HOME/.zsh/local" ]           || mkdir -p $HOME/.zsh/local
+[ -f "$HOME/.zsh/local/.zshrc" ]    || touch $HOME/.zsh/local/.zshrc
+[ -f "$HOME/.zsh/local/.zprofile" ] || touch $HOME/.zsh/local/.zprofile
